Cambios en la implementacion del publicador
