+++ clientesAPI/src/main/java/udea/clientesAPI/service/Publicador.java
@@ -2,6 +2,8 @@ package udea.clientesAPI.service;
 
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Service;
 import udea.clientesAPI.modelo.Cliente;
 
@@ -11,7 +13,12 @@ import java.util.concurrent.CompletableFuture;
  * Created by pabliny193@gmail.com on 05/12/2017.
  */
 @Service
+@PropertySource(value="classpath:application.properties")
+@ConfigurationProperties(prefix = "rabbit.exchange")
 public class Publicador {
+
+  private String nombreExchange;
+  private String routingKey;
 
   private final RabbitTemplate rabbitTemplate;
 
@@ -20,7 +27,23 @@ public class Publicador {
     this.rabbitTemplate = rabbitTemplate;
   }
 
-  public void publicarMensajeAsnc(String exchange, String routingK, Cliente cliente){
-    CompletableFuture.runAsync(()-> rabbitTemplate.convertAndSend(exchange, routingK, cliente));
+  public void publicarMensajeAsnc(Cliente cliente){
+    CompletableFuture.runAsync(()-> rabbitTemplate.convertAndSend(getNombreExchange(), getRoutingKey(), cliente));
+  }
+
+  public String getNombreExchange() {
+    return nombreExchange;
+  }
+
+  public void setNombreExchange(String nombreExchange) {
+    this.nombreExchange = nombreExchange;
+  }
+
+  public String getRoutingKey() {
+    return routingKey;
+  }
+
+  public void setRoutingKey(String routingKey) {
+    this.routingKey = routingKey;
   }
 }
