+++ clientesAPI/src/main/resources/application.properties
@@ -7,5 +7,5 @@ rabbit.configuration.username=vvyktpwm
 rabbit.configuration.password=qnTCT1lusVtLFXSyTDLmH0wA5xU0GxdF
 rabbit.configuration.virtualHost=vvyktpwm
 
-rabbit.queue.colaPublicador=udea.crm.retail.clientecreado
-rabbit.queue.routingKey=clientes
+rabbit.exchange.nombreExchange=udea.crm.retail.clientecreado
+rabbit.exchange.routingKey=clientes
